Move `Raw` into `raw` module
